+++ src/main/java/me/erykczy/colorfullighting/mixin/render/SpriteContentsAccessor.java
@@ -1,0 +1,12 @@
+package me.erykczy.colorfullighting.mixin.render;
+
+import com.mojang.blaze3d.platform.NativeImage;
+import net.minecraft.client.renderer.texture.SpriteContents;
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.gen.Accessor;
+
+@Mixin(SpriteContents.class)
+public interface SpriteContentsAccessor {
+    @Accessor("originalImage")
+    NativeImage colorfullighting$getOriginalImage();
+}